fixing the payment description and making answers

+++ models/staging/stg_green_tripdata.sql
@@ -37,7 +37,7 @@ select
     cast(tolls_amount as numeric) as tolls_amount,
     cast(improvement_surcharge as string) as improvement_surcharge,
     cast(total_amount as numeric) as total_amount,
-    coalesce({{ dbt.safe_cast("payment_type", api.Column.translate_type("integer")) }},0) as payment_type,
+    coalesce({{ dbt.safe_cast("payment_type", api.Column.translate_type("numeric")) }},0) as payment_type,
     {{ get_payment_type_description("payment_type") }} as payment_type_description
 from tripdata
 where rn = 1
